Added unit tests; Added Swagger documentation

+++ src/SoftServe_TestProject.API/Controllers/CoursesController.cs
@@ -3,8 +3,8 @@ using FluentValidation;
 using Microsoft.AspNetCore.Mvc;
 using SoftServe_TestProject.API.DTOs;
 using SoftServe_TestProject.API.Responses;
+using SoftServe_TestProject.Application.Interfaces;
 using SoftServe_TestProject.Application.Requests;
-using SoftServe_TestProject.Application.Services;
 
 namespace SoftServe_TestProject.API.Controllers
 {
@@ -12,24 +12,30 @@ namespace SoftServe_TestProject.API.Controllers
     [Route("api/courses")]
     public class CoursesController : ControllerBase
     {
-        private readonly CourseService _courseService;
+        private readonly ICourseService _courseService;
         private readonly IValidator<CourseDTO> _courseValidator;
         private readonly IMapper _mapper;
 
-        public CoursesController(CourseService courseService, IValidator<CourseDTO> courseValidator, IMapper mapper)
+        public CoursesController(ICourseService courseService, IValidator<CourseDTO> courseValidator, IMapper mapper)
         {
             _courseService = courseService;
             _courseValidator = courseValidator;
             _mapper = mapper;
         }
 
+        /// <summary>
+        /// Gets the course by id
+        /// </summary>
+        /// <returns>The course entity</returns>
+        /// <response code="200">Returns the course entity</response>
+        /// <response code="404">The entity was not found</response>
         [HttpGet("{id:int}")]
         public async Task<IActionResult> GetById(int id)
         {
             var course = await _courseService.GetCourseByIdAsync(id);
             if (course == null)
             {
-                return NotFound("Course not found.");
+                return NotFound(new { Error = "Course not found." } );
             }
 
             var courseResponse = _mapper.Map<CourseResponse>(course);
@@ -37,6 +43,11 @@ namespace SoftServe_TestProject.API.Controllers
             return Ok(courseResponse);
         }
 
+        /// <summary>
+        /// Gets the list of all courses
+        /// </summary>
+        /// <returns>The list of courses</returns>
+        /// <response code="200">Returns the list of courses</response>
         [HttpGet]
         public async Task<IActionResult> GetAll()
         {
@@ -47,13 +58,19 @@ namespace SoftServe_TestProject.API.Controllers
             return Ok(courseResponses);
         }
 
+        /// <summary>
+        /// Creates the new course
+        /// </summary>
+        /// <returns>No content</returns>
+        /// <response code="204">The course was successfully created</response>
+        /// <response code="400">The entity was not valid</response>
         [HttpPost]
         public async Task<IActionResult> Create(CourseDTO courseDTO)
         {
             var validationResult = await _courseValidator.ValidateAsync(courseDTO);
             if (!validationResult.IsValid)
             {
-                return BadRequest(new { Errors = validationResult.Errors.Select(e => e.ErrorMessage) });
+                return BadRequest(new { Error = validationResult.Errors.Select(e => e.ErrorMessage) });
             }
 
             var course = _mapper.Map<CourseRequest>(courseDTO);
@@ -62,6 +79,13 @@ namespace SoftServe_TestProject.API.Controllers
             return NoContent();
         }
 
+        /// <summary>
+        /// Updates the existing course
+        /// </summary>
+        /// <returns>No content</returns>
+        /// <response code="204">The course was successfully updated</response>
+        /// <response code="400">The course was not valid</response>
+        /// <response code="400">Id and course's id were different</response>
         [HttpPut("{id:int}")]
         public async Task<IActionResult> Update(int id, CourseDTO courseDTO)
         {
@@ -73,7 +97,7 @@ namespace SoftServe_TestProject.API.Controllers
             var validationResult = await _courseValidator.ValidateAsync(courseDTO);
             if (!validationResult.IsValid)
             {
-                return BadRequest(new { Errors = validationResult.Errors.Select(e => e.ErrorMessage) });
+                return BadRequest(new { Error = validationResult.Errors.Select(e => e.ErrorMessage) });
             }
 
             var course = _mapper.Map<CourseRequest>(courseDTO);
@@ -82,9 +106,21 @@ namespace SoftServe_TestProject.API.Controllers
             return NoContent();
         }
 
+        /// <summary>
+        /// Deletes the course by id
+        /// </summary>
+        /// <returns>No content</returns>
+        /// <response code="204">The course was successfully deleted</response>
+        /// <response code="404">The entity was not found</response>
         [HttpDelete("{id:int}")]
         public async Task<IActionResult> Delete(int id)
         {
+            var course = await _courseService.GetCourseByIdAsync(id);
+            if (course == null)
+            {
+                return NotFound(new { Error = "Course not found." } );
+            }
+
             await _courseService.DeleteCourseAsync(id);
 
             return NoContent();
